[Justice Counts] Refactor to improve performance (Recidiviz/recidiviz-data#15619)
## Description of the change

This PR does a few things to improve performance:
1) The biggest change is removing our calls to `get_existing_entity` and
`update_existing_or_create`, which are expensive database lookups.
Instead, we pass around an `existing_datapoints_dict` and perform
lookups against that dictionary instead.
2) We also reduce the number of calls to `get_reports_by_agency_id` and
`update_report_metadata`.

This refactor should be a no-op, so existing tests passing is a good
sign that it works. I'm also testing rigorously both locally and on
staging!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ffd19892236a9a148b4abbcb67b2b817f83edb43