[Workflows] Add given_name and surname to staff_record view and transformation (Recidiviz/recidiviz-data#15615)
This PR updates the supervision_staff_template views per state and staff_record parent view to include given_names and surname columns. It also includes the change to transform these fields in the WorkflowsStaffETLDelegate. This will be used to unblock sorting officers by name. Tested by exporting the result of the updated staff_record query to the US_ND bucket and manually running the ETL.

GitOrigin-RevId: 7064ffe9b20400d943ed055152b3811566c6c59d