[Workflows] Add generalized client record ETL delegate (Recidiviz/recidiviz-data#15553)
## Description of the change

This PR adds the `WorkflowsClientETLDelegate` which will replace the
state specific delegates. There is also a new file,
`state_specific_etl_transformations`.

I included the fields from LSU and ED because I originally branched off
of that client record PR, but then realized I didn't want this to be
held up by that work, so I rebased onto main. I don't think there would
be any issues with including the eligibility fields in the delegate
while that work is still ongoing, but let me know if I'm missing
anything and should remove it.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#14727

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 44450fcb5d493ff9a3ec8db07b4e5d2b69e3cd69