[Justice Counts] Change metric unique keys so they are not dependent on disaggregations (Recidiviz/recidiviz-data#15663)
## Description of the change

Our MetricDefinition key property what we use to uniquely identify an
official Justice Counts metric. It is
[calculated](https://github.com/Recidiviz/recidiviz-data/blob/f2e9cdb0c93761769d2436e47871159f865847cf/recidiviz/justice_counts/metrics/metric_definition.py#L159)
to be equal to the system + metric type + list of aggregation
dimensions, for example
`LAW_ENFORCEMENT_REPORTED_CRIME_metric/law_enforcement/reported_crime/type`
represents a metric for the Law Enforcement system, with type
"REPORTED_CRIME", with a disaggregation on "crime type".

These keys are [stored in our database on each
Datapoint](https://github.com/Recidiviz/recidiviz-data/blob/44450fcb5d493ff9a3ec8db07b4e5d2b69e3cd69/recidiviz/persistence/database/schema/justice_counts/schema.py#L501).
It's how we know which Justice Counts metric this datapoint belongs to.
This means that if the key for a MetricDefinition changes, these
datapoints become "orphans" -- when we try to look up what metric they
belong to, we don't find the matching one.

This is a problem if we ever need to adjust the disaggregations on a
metric. In the past, when this happened, we had to run a
[migration](https://github.com/Recidiviz/recidiviz-data/pull/14492) to
change the keys in the database to "unorphan" the datapoints. But this
is not sustainable moving forwards, so we need to figure out a better
long-term solution. This is especially important because soon we're
going to be adjusting a lot of MetricDefinitions to match the latest
updates to the Justice Counts metrics.

I think the right fix is to have key equal system + metric type ONLY, so
e.g. we'd just have `LAW_ENFORCEMENT_REPORTED_CRIME`. In order to make
this work, we have to take the following steps:

First, we need to make sure that system + metric type uniquely
identifies all metrics. In other words, are there any MetricDefinitions
that have the same system and metric type, and only differ in their
aggregations? If so, we need to create a new metric type for one of
them. This is done in this PR (i.e. we started to use SUPERVISION_STARTS
metric type instead of using POPULATION twice).

Second, we need to run a BIG database migration, like the one
[here](https://github.com/Recidiviz/recidiviz-data/pull/14492/files),
but now to change every single metric key!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#15407

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 80318bb2565ade97def038e18e4f273d4831dade