[Metrics] Fixes supervision population definition, fills known zeros (Recidiviz/recidiviz-data#15608)
## Description of the change

Two fixes:

1. Fixes the supervision population definition - must have an
intersection of assignment to officer/office/district/state _and_ be in
one of the compartment_level_2s of interest (parole, probation, etc.)
2. `IFNULL`s all `sum` metrics to zeros where a null value indicates
zero. Leaves other metrics that can be null as null, and adds comments
to explain.

I copied sandbox tables to the `unamanaged_metrics` dataset, which we
are currently referencing in Looker (because of a state-partner
presentation).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f5e01dba1a740465a0225a8dc1a5591a9ab46dd0