Update issue templates (Recidiviz/recidiviz-data#15604)
## Description of the change

This PR slightly modifies some GitHub Issue templates to migrate the
`State Pod: XXXX` labels to `Team: State Pod`. This snippet prevents the
text from appearing in the rendered ticket, but it is still visible in
the raw markdown. Other spacing/formatting issues have also been
addressed.

~It also adds the following snippet `[//]:# ` as a prefix before text
that is useful for the ticket-writer, but can add clutter to the ticket
after it has been created.~ This part wasn't working out properly, but
would be nice to have if anyone can figure it out (see
[this](https://github.com/Recidiviz/recidiviz-research/blob/858d9f7d1160b6e614fcc11d494d5dad17fe7bc4/.github/ISSUE_TEMPLATE/research_task.md)
template as an example).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b3a02a2e59d2ace929772ad62d66cc09956e2c8f