[US_MI] Add supervision population external validation (Recidiviz/recidiviz-data#15498)
## Description of the change

Add supervision population by type external validation.

For MI, the external validation data comes from the
Parole/Probation/Non-Resident section of the CB971 report. Even though
this report is run once a week (and we use other sections of the report
for the incarceration population external validation), the population
counts for the Parole/Probation/Non-Resident section are only updated
once a month sometime around the middle of the month. Therefore, for
this validation for MI, I will be using the report from the third week
of each month for it to be closest to when the numbers of accurate for.
However, since we won't have an exact idea of when the counts are
accurate for, we'll probably want to bump up the error threshold. In
addition, the external validation data only contains counts for parole
and probation separately (and therefore would not account for dual
supervision either).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 04aab9c23d066989e33b356f3cb61d5a5eb1d5c3