[Justice Counts] If datapoint history is an invalid number, keep type as string. (Recidiviz/recidiviz-data#15709)
## Description of the change

This PR includes a fix
[this](https://console.cloud.google.com/monitoring/alerting/policies/2585565922764144568?project=recidiviz-123)
bug that came up in production. This bug can be replicated by adding an
invalid value (i.e $4,123,231) to a report via the data entry page and
then changing it such that it is no longer invalid (i.e $4,123,231 ->
4,123,231). The `cast_to_number` function was failing to cast the
invalid datapoint history value as an int.

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1457d83d6da5cf22917fcb02162186848b19c976